perf: optimize production bundle size with code splitting and vendor chunking
- Split MainApp and PublicSharePage into separate lazy-loaded components.
- Configure Vite manual chunks for Firebase vendor libraries.
- Centralize shared utility functions to improve maintainability and tree-shaking.
- Main entry chunk size reduced from ~590kB to ~193kB.
